x64 port
git-svn-id: http://miranda.googlecode.com/svn/trunk@9308 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/protocols/JabberG/jabber_iqid.cpp
@@ -773,7 +773,7 @@ void CJabberProto::OnIqResultGetVcard( HXML iqNode )
 		return;
 	}
 
-	int len = _tcslen( m_szJabberJID );
+	size_t len = _tcslen( m_szJabberJID );
 	if ( !_tcsnicmp( jid, m_szJabberJID, len ) && ( jid[len]=='/' || jid[len]=='\0' )) {
 		hContact = NULL;
 		Log( "Vcard for myself" );
